checkpoint: renames "TextToImage/client" to "TextToImage/Client"
