+++ src/features/TextToImage/Client/SDXL/conversions/GenerateFromTextResponse.ts
