+++ src/features/TextToImage/Client/SDXL/conversions/StabilityAI_Client_Image.ts
